# services/app-auth/internal/dto/http.gen.go
// Package dto provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.13.0 DO NOT EDIT.
package dto

import (
	"github.com/MikeMwita/fedha.git/services/app-auth/internal/core/entity"
	"time"

	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

// CreateExpenseTypeRequest defines model for CreateExpenseTypeRequest.
type CreateExpenseTypeRequest struct {
	Description *string `json:"description,omitempty"`
	Name        string  `json:"name"`
}

// ExpenseTypeResponse defines model for ExpenseTypeResponse.
type ExpenseTypeResponse struct {
	Description   *string `json:"description,omitempty"`
	ExpenseTypeId *int    `json:"expense_type_id,omitempty"`
	Name          *string `json:"name,omitempty"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// LoginResponseData defines model for LoginResponseData.
type LoginResponseData struct {
	UserId                entity.User
	*string               `json:"access_token,omitempty"`
	AccessTokenExpiresAt  *time.Time            `json:"access_token_expires_at,omitempty"`
	RefreshToken          *string               `json:"refresh_token,omitempty"`
	RefreshTokenExpiresAt *time.Time            `json:"refresh_token_expires_at,omitempty"`
	SessionId             *openapi_types.UUID   `json:"session_id,omitempty"`
	User                  *RegisterResponseData `json:"user,omitempty"`
}

// LogoutResponse defines model for LogoutResponse.
type LogoutResponse struct {
	Message string `json:"message"`
}

// RefreshTokenRequest defines model for RefreshTokenRequest.
type RefreshTokenRequest struct {
	RefreshToken string `json:"refresh_token"`
}

// RefreshTokenResponse defines model for RefreshTokenResponse.
type RefreshTokenResponse struct {
	AccessToken string `json:"access_token"`
	ExpiresIn   string `json:"expires_in"`
}

// RegisterRequest defines model for RegisterRequest.
type RegisterRequest struct {
	Email    openapi_types.Email `json:"email"`
	FullName string              `json:"full_name"`
	Password string              `json:"password"`
	Username string              `json:"username"`
}



// RegisterResponseData defines model for RegisterResponseData.
type RegisterResponseData struct {
	CreatedAt         *time.Time `json:"created_at,omitempty"`
	Email             *string    `json:"email,omitempty"`
	FullName          *string    `json:"full_name,omitempty"`
	PasswordChangedAt *time.Time `json:"password_changed_at,omitempty"`
	Username          *string    `json:"username,omitempty"`
}

// Response defines model for Response.
type Response struct {
	Error   *string `json:"error,omitempty"`
	Message *string `json:"message,omitempty"`
}

// CreateExpenseTypeJSONRequestBody defines body for CreateExpenseType for application/json ContentType.
type CreateExpenseTypeJSONRequestBody = CreateExpenseTypeRequest

// RefreshTokenJSONRequestBody defines body for RefreshToken for application/json ContentType.
type RefreshTokenJSONRequestBody = RefreshTokenRequest

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// RegisterJSONRequestBody defines body for Register for application/json ContentType.
type RegisterJSONRequestBody = RegisterRequest